scripts/requirements.txt
azure-ai-documentintelligence==1.0.2
python-dotenv